refactor: move NetworkException to core.network package
NetworkException belongs in the top-level network package, not the
retrofit subpackage, since Retrofit has been removed.
