+++ core/network/src/main/kotlin/com/merxury/blocker/core/network/retrofit/OkHttpBlockerNetwork.kt
@@ -20,6 +20,7 @@ package com.merxury.blocker.core.network.retrofit
 import com.merxury.blocker.core.model.preference.RuleServerProvider
 import com.merxury.blocker.core.model.preference.RuleServerProvider.GITHUB
 import com.merxury.blocker.core.network.BlockerNetworkDataSource
+import com.merxury.blocker.core.network.NetworkException
 import com.merxury.blocker.core.network.io.BinaryFileWriter
 import com.merxury.blocker.core.network.model.NetworkChangeList
 import kotlinx.coroutines.Dispatchers
